@@ -205,26 +135,20 @@ else()
   endif()
 endif()
 
-# Create a runtime.
-add_custom_command(
-  OUTPUT runtime_32.o
-  COMMAND "${CMAKE_BC_COMPILER}" -std=gnu++11 -isystem "${REMILL_SOURCE_DIR}" -isystem "${MCSEMA_SOURCE_DIR}" -m32 -fPIC -c "${CMAKE_CURRENT_SOURCE_DIR}/Runtime.cpp" -o runtime_32.o
-  DEPENDS Runtime.cpp
-  COMMENT "Building 32-bit runtime"
-)
 
-set_source_files_properties(
-  runtime_32.o
-  PROPERTIES
-  EXTERNAL_OBJECT True
-  GENERATED True
-)
+if(DEFINED REMILL_INCLUDE_LOCATION)
+  set(REMILL_INCLUDE "-I${REMILL_INCLUDE_LOCATION}")
+endif()
+
+if(DEFINED REMILL_SOURCE_DIR)
+  set(REMILL_SOURCE "-I${REMILL_SOURCE_DIR}")
+endif()
 
 # only do 64-bit binaries on 64-bit platforms
 if(CMAKE_SIZEOF_VOID_P EQUAL 8)
   add_custom_command(
     OUTPUT runtime_64.o
-    COMMAND "${CMAKE_BC_COMPILER}" -std=gnu++11 -isystem "${REMILL_SOURCE_DIR}" -isystem "${MCSEMA_SOURCE_DIR}" -m64  -fPIC -c "${CMAKE_CURRENT_SOURCE_DIR}/Runtime.cpp" -o runtime_64.o
+    COMMAND "${CMAKE_CXX_COMPILER}" -std=gnu++11 "${REMILL_INCLUDE}" "${REMILL_SOURCE}" -I"${MCSEMA_SOURCE_DIR}" -m64  -fPIC -c "${CMAKE_CURRENT_SOURCE_DIR}/Runtime.cpp" -o runtime_64.o
     DEPENDS Runtime.cpp
     COMMENT "Building 64-bit runtime"
   )
@@ -234,5 +158,15 @@ if(CMAKE_SIZEOF_VOID_P EQUAL 8)
     PROPERTIES
     EXTERNAL_OBJECT True
     GENERATED True
+  )
+endif()
+
+if(CMAKE_SIZEOF_VOID_P EQUAL 8)
+  add_runtime(runtime_amd64
+    SOURCES "Runtime.cpp"
+    ADDRESS_SIZE 64
+    BCFLAGS  "-xc++" "-m64" "-std=gnu++17" "-Wno-deprecated-declarations"
+    INCLUDEDIRECTORIES "${CMAKE_SOURCE_DIR}" "${MCSEMA_SOURCE_DIR}"  "${REMILL_INCLUDE_LOCATION}"
+    INSTALLDESTINATION "${install_folder}/lib"
   )
 endif()
