@@ -1,37 +1,40 @@
 /*
- * Copyright (c) 2017 Trail of Bits, Inc.
+ * Copyright (c) 2020 Trail of Bits, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <cstdio>
 #include <cinttypes>
+#include <cstdio>
 
 #define HAS_FEATURE_AVX 1
 #define HAS_FEATURE_AVX512 0
 #define ADDRESS_SIZE_BITS 32
 
-#include <remill/Arch/X86/Runtime/State.h>
 #include <mcsema/Arch/X86/Runtime/Registers.h>
+#include <remill/Arch/X86/Runtime/State.h>
 
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wformat"
 
 static const size_t kStackSize = 1UL << 20UL;
 
-static void PrintStoreFlags(FILE * out) {
+static void PrintStoreFlags(FILE *out) {
+
   // FPU control.
-  fprintf(out, "  fnstcw WORD PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, x87.fxsave.cwd));
+  fprintf(out, "  fnstcw WORD PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, x87.fxsave.cwd));
 
   fprintf(out, "  pushfd\n");
   fprintf(out, "  mov edx, 0xcd5\n");
@@ -39,81 +42,102 @@ static void PrintStoreFlags(FILE * out) {
   fprintf(out, "  and DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, CF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, CF));
   fprintf(out, "  shl edx, 0\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, PF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, PF));
   fprintf(out, "  shl edx, 2\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, AF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, AF));
   fprintf(out, "  shl edx, 4\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ZF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ZF));
   fprintf(out, "  shl edx, 6\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, SF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, SF));
   fprintf(out, "  shl edx, 7\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, DF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, DF));
   fprintf(out, "  shl edx, 10\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, OF));
+  fprintf(out, "  and dl, BYTE PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, OF));
   fprintf(out, "  shl edx, 11\n");
   fprintf(out, "  or DWORD PTR [esp], edx\n");
 
   fprintf(out, "  popfd\n");
 }
 
-static void PrintLoadFlags(FILE * out) {
+static void PrintLoadFlags(FILE *out) {
+
   // FPU control.
   fprintf(out, "  push dx\n");
   fprintf(out, "  fldcw WORD PTR [esp]\n");
-  fprintf(out, "  pop WORD PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, x87.fxsave.cwd));
+  fprintf(out, "  pop WORD PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, x87.fxsave.cwd));
 
   // Clear our the `ArithFlags` struct, which is 16 bytes.
-  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, aflag));
-  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, aflag) + 4);
-  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, aflag) + 8);
-  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, aflag) + 12);
+  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, aflag));
+  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, aflag) + 4);
+  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, aflag) + 8);
+  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, aflag) + 12);
 
   // Get the EFlags.
   fprintf(out, "  pushfd\n");
   fprintf(out, "  pop edx\n");
-  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], edx\n", __builtin_offsetof(State, rflag));
+  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], edx\n",
+          __builtin_offsetof(State, rflag));
 
   // Marshal the EFlags into the ArithFlags struct.
   fprintf(out, "  bt edx, 0\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, CF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, CF));
 
   fprintf(out, "  bt edx, 2\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, PF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, PF));
 
   fprintf(out, "  bt edx, 4\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, AF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, AF));
 
   fprintf(out, "  bt edx, 6\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, ZF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, ZF));
 
   fprintf(out, "  bt edx, 7\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, SF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, SF));
 
   fprintf(out, "  bt edx, 10\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, DF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, DF));
 
   fprintf(out, "  bt edx, 11\n");
-  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, OF));
+  fprintf(out, "  adc BYTE PTR [edi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, OF));
 }
 
 int main(void) {
@@ -158,8 +182,10 @@ int main(void) {
   fprintf(out, "  .cfi_startproc\n");
 
   // General-purpose registers.
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n", __builtin_offsetof(State, ESI));
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n", __builtin_offsetof(State, EDI));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n",
+          __builtin_offsetof(State, ESI));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n",
+          __builtin_offsetof(State, EDI));
 
   // Set up the `GS` segment register so that TLS works :-)
   fprintf(out, "  mov esi, DWORD PTR gs:[0]\n");
@@ -172,32 +198,47 @@ int main(void) {
   // `EDI` points to the State structure
 
   // Rest of the GPRs.
-  fprintf(out, "  mov [edi + %" PRIuMAX "], edx\n", __builtin_offsetof(State, EDX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], eax\n", __builtin_offsetof(State, EAX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ebx\n", __builtin_offsetof(State, EBX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ecx\n", __builtin_offsetof(State, ECX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ebp\n", __builtin_offsetof(State, EBP));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], edx\n",
+          __builtin_offsetof(State, EDX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], eax\n",
+          __builtin_offsetof(State, EAX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ebx\n",
+          __builtin_offsetof(State, EBX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ecx\n",
+          __builtin_offsetof(State, ECX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ebp\n",
+          __builtin_offsetof(State, EBP));
 
   // XMM registers.
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm0\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm1\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm2\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm3\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm4\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm5\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm6\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm7\n", __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm0\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm1\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm2\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm3\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm4\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm5\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm6\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm7\n",
+          __builtin_offsetof(State, XMM7));
 
   PrintLoadFlags(out);  // Note: Clobbers EDX.
 
   // If `ESP` is null then we need to initialize it to our new stack.
   //
   // Note: `ESI` is the TLS base.
-  fprintf(out, "  mov edx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  mov edx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ESP));
   fprintf(out, "  cmp edx, 0\n");
   fprintf(out, "  jnz .Lhave_stack\n");
   fprintf(out, "  lea eax, DWORD PTR [__mcsema_stack@TPOFF]\n");
-  fprintf(out, "  lea edx, DWORD PTR [esi + eax + %" PRIuMAX "]\n", (kStackSize - 16));
+  fprintf(out, "  lea edx, DWORD PTR [esi + eax + %" PRIuMAX "]\n",
+          (kStackSize - 16));
   fprintf(out, ".Lhave_stack:\n");
 
   // `ESI` is the TLS base
@@ -211,7 +252,8 @@ int main(void) {
 
   // Get the program counter off of the stack.
   fprintf(out, "  pop esi\n");
-  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], esi\n", __builtin_offsetof(State, EIP));
+  fprintf(out, "  mov DWORD PTR [edi + %" PRIuMAX "], esi\n",
+          __builtin_offsetof(State, EIP));
 
   // Set up the arguments on the McSema stack: state, pc, memory.
   fprintf(out, "  lea edx, [edx - 20]\n");
@@ -228,14 +270,16 @@ int main(void) {
   fprintf(out, "  pop DWORD PTR [edx]\n", __builtin_offsetof(State, EIP));
 
   // Swap onto the lifted stack. The native `ESP` is now where it should be.
-  fprintf(out, "  mov [edi + %" PRIuMAX "], esp\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], esp\n",
+          __builtin_offsetof(State, ESP));
   fprintf(out, "  mov esp, edx\n");
 
   // Ret to the target.
   fprintf(out, "  ret\n");
 
   fprintf(out, ".Lfunc_end1:\n");
-  fprintf(out, "  .size __mcsema_attach_call,.Lfunc_end1-__mcsema_attach_call\n");
+  fprintf(out,
+          "  .size __mcsema_attach_call,.Lfunc_end1-__mcsema_attach_call\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -259,30 +303,47 @@ int main(void) {
   // We "undo" that, then swap back to the native stack. When we swap, we
   // save into `State::RSP` where we are in the lifted stack, so that the
   // next attach can continue on where we left off.
-  fprintf(out, "  sub DWORD PTR [edi + %" PRIuMAX "], 4\n", __builtin_offsetof(State, ESP));
-  fprintf(out, "  xchg [edi + %" PRIuMAX "], esp\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  sub DWORD PTR [edi + %" PRIuMAX "], 4\n",
+          __builtin_offsetof(State, ESP));
+  fprintf(out, "  xchg [edi + %" PRIuMAX "], esp\n",
+          __builtin_offsetof(State, ESP));
 
   PrintStoreFlags(out);  // Clobbers RDX.
 
   // General purpose registers.
-  fprintf(out, "  mov eax, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EAX));
-  fprintf(out, "  mov ebx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EBX));
-  fprintf(out, "  mov ecx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ECX));
-  fprintf(out, "  mov edx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EDX));
-  fprintf(out, "  mov esi, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ESI));
-  fprintf(out, "  mov ebp, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EBP));
+  fprintf(out, "  mov eax, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EAX));
+  fprintf(out, "  mov ebx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EBX));
+  fprintf(out, "  mov ecx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ECX));
+  fprintf(out, "  mov edx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EDX));
+  fprintf(out, "  mov esi, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ESI));
+  fprintf(out, "  mov ebp, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EBP));
 
   // XMM registers.
-  fprintf(out, "  movntdqa xmm0, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdqa xmm1, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdqa xmm2, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdqa xmm3, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdqa xmm4, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdqa xmm5, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdqa xmm6, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdqa xmm7, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdqa xmm0, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdqa xmm1, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdqa xmm2, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdqa xmm3, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdqa xmm4, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdqa xmm5, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdqa xmm6, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdqa xmm7, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM7));
 
-  fprintf(out, "  mov edi, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EDI));
+  fprintf(out, "  mov edi, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EDI));
   fprintf(out, "  ret\n");
 
   fprintf(out, ".Lfunc_end3:\n");
@@ -330,7 +391,8 @@ int main(void) {
 
   // Stash the return address stored on the native stack, the replace it
   // with the re-attach function.
-  fprintf(out, "  mov esi, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  mov esi, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ESP));
   fprintf(out, "  push DWORD PTR [esi]\n");
   fprintf(out, "  push __mcsema_attach_ret\n");
   fprintf(out, "  pop DWORD PTR [esi]\n");
@@ -347,29 +409,45 @@ int main(void) {
   PrintStoreFlags(out);  // Clobbers EDX.
 
   // Swap off-stack, stash the lifted stack pointer.
-  fprintf(out, "  mov [edi + %" PRIuMAX "], esp\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], esp\n",
+          __builtin_offsetof(State, ESP));
   fprintf(out, "  mov esp, esi\n");
 
   // (Most) General purpose registers.
-  fprintf(out, "  mov eax, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EAX));
-  fprintf(out, "  mov ebx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EBX));
-  fprintf(out, "  mov ecx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ECX));
-  fprintf(out, "  mov edx, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EDX));
-  fprintf(out, "  mov esi, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ESI));
-  fprintf(out, "  mov ebp, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EBP));
+  fprintf(out, "  mov eax, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EAX));
+  fprintf(out, "  mov ebx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EBX));
+  fprintf(out, "  mov ecx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ECX));
+  fprintf(out, "  mov edx, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EDX));
+  fprintf(out, "  mov esi, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ESI));
+  fprintf(out, "  mov ebp, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EBP));
 
   // XMM registers.
-  fprintf(out, "  movntdqa xmm0, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdqa xmm1, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdqa xmm2, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdqa xmm3, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdqa xmm4, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdqa xmm5, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdqa xmm6, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdqa xmm7, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdqa xmm0, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdqa xmm1, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdqa xmm2, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdqa xmm3, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdqa xmm4, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdqa xmm5, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdqa xmm6, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdqa xmm7, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM7));
 
   // Swap out EDI.
-  fprintf(out, "  mov edi, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EDI));
+  fprintf(out, "  mov edi, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EDI));
 
   // Code above put the native target address (stored in RDI on entry to
   // `__remill_function_call`) on the stack, just below the return address,
@@ -378,7 +456,9 @@ int main(void) {
   fprintf(out, "  ret\n");
 
   fprintf(out, ".Lfunc_end5:\n");
-  fprintf(out, "  .size __remill_function_call,.Lfunc_end5-__remill_function_call\n");
+  fprintf(
+      out,
+      "  .size __remill_function_call,.Lfunc_end5-__remill_function_call\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -392,33 +472,50 @@ int main(void) {
   // Copy RSI, then store the address of the reg state struct into RSI for
   // easier indexing later on. Also set up the `FS` segment register so that
   // TLS works :-)
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n", __builtin_offsetof(State, EDI));
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n", __builtin_offsetof(State, ESI));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n",
+          __builtin_offsetof(State, EDI));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n",
+          __builtin_offsetof(State, ESI));
 
   fprintf(out, "  mov edi, DWORD PTR gs:[0]\n");
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n", __builtin_offsetof(State, FS_BASE));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n",
+          __builtin_offsetof(State, FS_BASE));
   fprintf(out, "  lea esi, [__mcsema_reg_state@TPOFF]\n");
   fprintf(out, "  lea edi, DWORD PTR [esi + edi]\n");
 
   // General purpose registers.
-  fprintf(out, "  mov [edi + %" PRIuMAX "], eax\n", __builtin_offsetof(State, EAX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ebx\n", __builtin_offsetof(State, EBX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ecx\n", __builtin_offsetof(State, ECX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], edx\n", __builtin_offsetof(State, EDX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ebp\n", __builtin_offsetof(State, EBP));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], eax\n",
+          __builtin_offsetof(State, EAX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ebx\n",
+          __builtin_offsetof(State, EBX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ecx\n",
+          __builtin_offsetof(State, ECX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], edx\n",
+          __builtin_offsetof(State, EDX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ebp\n",
+          __builtin_offsetof(State, EBP));
 
   // Swap into the mcsema stack.
-  fprintf(out, "  xchg esp, [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  xchg esp, [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ESP));
 
   // XMM registers.
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm0\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm1\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm2\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm3\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm4\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm5\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm6\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm7\n", __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm0\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm1\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm2\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm3\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm4\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm5\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm6\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm7\n",
+          __builtin_offsetof(State, XMM7));
 
   PrintLoadFlags(out);  // Note: Clobbers EDX.
 
@@ -430,7 +527,8 @@ int main(void) {
   //    16    stashed ebx
 
   // Restore emulated return address.
-  fprintf(out, "  pop DWORD PTR [edi + %" PRIuMAX "]\n", __builtin_offsetof(State, EIP));
+  fprintf(out, "  pop DWORD PTR [edi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EIP));
 
 
   // Callee-saved registers.
@@ -458,44 +556,64 @@ int main(void) {
   fprintf(out, ".Lfunc_begin10:\n");
   fprintf(out, ".cfi_startproc\n");
 
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n", __builtin_offsetof(State, EDI));
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n", __builtin_offsetof(State, ESI));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n",
+          __builtin_offsetof(State, EDI));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n",
+          __builtin_offsetof(State, ESI));
 
   // Follow the cdecl ABI calling convention
   // Make `EDI` point to the stack pointer.
   fprintf(out, "  mov edi, DWORD PTR [esp + 24]\n");
+
   // Make `ESI` point to the frame pointer.
   fprintf(out, "  mov esi, DWORD PTR [esp + 28]\n");
+
   // Sets the native stack and base pointers
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n", __builtin_offsetof(State, ESP));
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n", __builtin_offsetof(State, EBP));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n",
+          __builtin_offsetof(State, ESP));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], esi\n",
+          __builtin_offsetof(State, EBP));
 
 
   fprintf(out, "  mov edi, DWORD PTR gs:[0]\n");
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n", __builtin_offsetof(State, FS_BASE));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], edi\n",
+          __builtin_offsetof(State, FS_BASE));
   fprintf(out, "  lea esi, [__mcsema_reg_state@TPOFF]\n");
   fprintf(out, "  lea edi, DWORD PTR [esi + edi]\n");
 
   // General purpose registers.
   //fprintf(out, "  mov [edi + %" PRIuMAX "], eax\n", __builtin_offsetof(State, EAX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ebx\n", __builtin_offsetof(State, EBX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], ecx\n", __builtin_offsetof(State, ECX));
-  fprintf(out, "  mov [edi + %" PRIuMAX "], edx\n", __builtin_offsetof(State, EDX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ebx\n",
+          __builtin_offsetof(State, EBX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], ecx\n",
+          __builtin_offsetof(State, ECX));
+  fprintf(out, "  mov [edi + %" PRIuMAX "], edx\n",
+          __builtin_offsetof(State, EDX));
 
   // XMM registers.
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm0\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm1\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm2\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm3\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm4\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm5\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm6\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm7\n", __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm0\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm1\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm2\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm3\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm4\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm5\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm6\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdq [edi + %" PRIuMAX "], xmm7\n",
+          __builtin_offsetof(State, XMM7));
 
   fprintf(out, "  ret\n");
   fprintf(out, "  ud2\n");
   fprintf(out, ".Lfunc_end10:\n");
-  fprintf(out, "  .size __mcsema_exception_ret,.Lfunc_end10-__mcsema_exception_ret\n");
+  fprintf(
+      out,
+      "  .size __mcsema_exception_ret,.Lfunc_end10-__mcsema_exception_ret\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -504,10 +622,13 @@ int main(void) {
   fprintf(out, "  .type __mcsema_get_stack_pointer,@function\n");
   fprintf(out, "__mcsema_get_stack_pointer:\n");
   fprintf(out, "  .cfi_startproc\n");
-  fprintf(out, "  mov eax, gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n", __builtin_offsetof(State, ESP));
+  fprintf(out, "  mov eax, gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ESP));
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end20:\n");
-  fprintf(out, "  .size __mcsema_get_stack_pointer,.Lfunc_end20-__mcsema_get_stack_pointer\n");
+  fprintf(
+      out,
+      "  .size __mcsema_get_stack_pointer,.Lfunc_end20-__mcsema_get_stack_pointer\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -516,10 +637,13 @@ int main(void) {
   fprintf(out, "  .type __mcsema_get_frame_pointer,@function\n");
   fprintf(out, "__mcsema_get_frame_pointer:\n");
   fprintf(out, "  .cfi_startproc\n");
-  fprintf(out, "  mov eax, gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n", __builtin_offsetof(State, EBP));
+  fprintf(out, "  mov eax, gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, EBP));
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end21:\n");
-  fprintf(out, "  .size __mcsema_get_frame_pointer,.Lfunc_end21-__mcsema_get_frame_pointer\n");
+  fprintf(
+      out,
+      "  .size __mcsema_get_frame_pointer,.Lfunc_end21-__mcsema_get_frame_pointer\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -528,11 +652,14 @@ int main(void) {
   fprintf(out, "  .type __mcsema_get_type_index,@function\n");
   fprintf(out, "__mcsema_get_type_index:\n");
   fprintf(out, "  .cfi_startproc\n");
-  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], eax\n", __builtin_offsetof(State, EAX));
+  fprintf(out, "  mov gs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], eax\n",
+          __builtin_offsetof(State, EAX));
   fprintf(out, "  mov eax, edx\n");
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end22:\n");
-  fprintf(out, "  .size __mcsema_get_type_index,.Lfunc_end22-__mcsema_get_type_index\n");
+  fprintf(
+      out,
+      "  .size __mcsema_get_type_index,.Lfunc_end22-__mcsema_get_type_index\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -547,7 +674,9 @@ int main(void) {
   fprintf(out, "  lea eax, [eax + edx]\n");
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end6:\n");
-  fprintf(out, "  .size __mcsema_debug_get_reg_state,.Lfunc_end6-__mcsema_debug_get_reg_state\n");
+  fprintf(
+      out,
+      "  .size __mcsema_debug_get_reg_state,.Lfunc_end6-__mcsema_debug_get_reg_state\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -578,4 +707,3 @@ int main(void) {
 
 //  // Restore stack alignment
 //  fprintf(out, "  pop esp\n");
-
