@@ -1,37 +1,40 @@
 /*
- * Copyright (c) 2017 Trail of Bits, Inc.
+ * Copyright (c) 2020 Trail of Bits, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <cstdio>
 #include <cinttypes>
+#include <cstdio>
 
 #define HAS_FEATURE_AVX 1
 #define HAS_FEATURE_AVX512 0
 #define ADDRESS_SIZE_BITS 64
 
-#include <remill/Arch/X86/Runtime/State.h>
 #include <mcsema/Arch/X86/Runtime/Registers.h>
+#include <remill/Arch/X86/Runtime/State.h>
 
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wformat"
 
 static const size_t kStackSize = 1UL << 20UL;
 
-static void PrintStoreFlags(FILE * out) {
+static void PrintStoreFlags(FILE *out) {
+
   // FPU control.
-  fprintf(out, "  fnstcw WORD PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, x87.fxsave.cwd));
+  fprintf(out, "  fnstcw WORD PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, x87.fxsave.cwd));
 
   fprintf(out, "  pushfq\n");
   fprintf(out, "  mov edx, 0xcd5\n");
@@ -39,79 +42,98 @@ static void PrintStoreFlags(FILE * out) {
   fprintf(out, "  and QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, CF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, CF));
   fprintf(out, "  shl edx, 0\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, PF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, PF));
   fprintf(out, "  shl edx, 2\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, AF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, AF));
   fprintf(out, "  shl edx, 4\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, ZF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, ZF));
   fprintf(out, "  shl edx, 6\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, SF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, SF));
   fprintf(out, "  shl edx, 7\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, DF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, DF));
   fprintf(out, "  shl edx, 10\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  mov edx, 1\n");
-  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, OF));
+  fprintf(out, "  and dl, BYTE PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, OF));
   fprintf(out, "  shl edx, 11\n");
   fprintf(out, "  or QWORD PTR [rsp], rdx\n");
 
   fprintf(out, "  popfq\n");
 }
 
-static void PrintLoadFlags(FILE * out) {
+static void PrintLoadFlags(FILE *out) {
+
   // FPU control.
   fprintf(out, "  push dx\n");
   fprintf(out, "  fldcw WORD PTR [rsp]\n");
-  fprintf(out, "  pop WORD PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, x87.fxsave.cwd));
+  fprintf(out, "  pop WORD PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, x87.fxsave.cwd));
 
   // Get the RFlags.
   fprintf(out, "  pushfq\n");
   fprintf(out, "  pop rdx\n");
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rdx\n", __builtin_offsetof(State, rflag));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rdx\n",
+          __builtin_offsetof(State, rflag));
 
   // Clear our the `ArithFlags` struct, which is 16 bytes.
-  fprintf(out, "  mov QWORD PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, aflag));
-  fprintf(out, "  mov QWORD PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, aflag) + 8);
+  fprintf(out, "  mov QWORD PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, aflag));
+  fprintf(out, "  mov QWORD PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, aflag) + 8);
 
   // Marshal the RFlags into the ArithFlags struct.
   fprintf(out, "  bt rdx, 0\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, CF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, CF));
 
   fprintf(out, "  bt rdx, 2\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, PF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, PF));
 
   fprintf(out, "  bt rdx, 4\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, AF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, AF));
 
   fprintf(out, "  bt rdx, 6\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, ZF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, ZF));
 
   fprintf(out, "  bt rdx, 7\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, SF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, SF));
 
   fprintf(out, "  bt rdx, 10\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, DF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, DF));
 
   fprintf(out, "  bt rdx, 11\n");
-  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n", __builtin_offsetof(State, OF));
+  fprintf(out, "  adc BYTE PTR [rdi + %" PRIuMAX "], 0\n",
+          __builtin_offsetof(State, OF));
 }
 
 int main(void) {
@@ -156,9 +178,12 @@ int main(void) {
   fprintf(out, "  .cfi_startproc\n");
 
   // Save off the first three args of the ABI.
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n", __builtin_offsetof(State, RSI));
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n", __builtin_offsetof(State, RDI));
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdx\n", __builtin_offsetof(State, RDX));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n",
+          __builtin_offsetof(State, RSI));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n",
+          __builtin_offsetof(State, RDI));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdx\n",
+          __builtin_offsetof(State, RDX));
 
   // On the stack:
   //     0  EA of the lifted function (from the CFG).
@@ -167,56 +192,88 @@ int main(void) {
 
   // Set up the `FS` segment register so that TLS works :-)
   fprintf(out, "  mov rsi, QWORD PTR fs:[0]\n");
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n", __builtin_offsetof(State, FS_BASE));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n",
+          __builtin_offsetof(State, FS_BASE));
 
   // Get arg (rdi) to contain the State pointer.
   fprintf(out, "  lea rdi, QWORD PTR [__mcsema_reg_state@TPOFF]\n");
   fprintf(out, "  lea rdi, QWORD PTR [rsi + rdi]\n");
 
   // Get the program counter off of the stack.
-  fprintf(out, "  pop QWORD PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RIP));
+  fprintf(out, "  pop QWORD PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RIP));
 
   // Remaining general purpose registers.
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n", __builtin_offsetof(State, RAX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbx\n", __builtin_offsetof(State, RBX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rcx\n", __builtin_offsetof(State, RCX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbp\n", __builtin_offsetof(State, RBP));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r8\n", __builtin_offsetof(State, R8));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r9\n", __builtin_offsetof(State, R9));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r10\n", __builtin_offsetof(State, R10));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r11\n", __builtin_offsetof(State, R11));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r12\n", __builtin_offsetof(State, R12));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r13\n", __builtin_offsetof(State, R13));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r14\n", __builtin_offsetof(State, R14));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r15\n", __builtin_offsetof(State, R15));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n",
+          __builtin_offsetof(State, RAX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbx\n",
+          __builtin_offsetof(State, RBX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rcx\n",
+          __builtin_offsetof(State, RCX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbp\n",
+          __builtin_offsetof(State, RBP));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r8\n",
+          __builtin_offsetof(State, R8));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r9\n",
+          __builtin_offsetof(State, R9));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r10\n",
+          __builtin_offsetof(State, R10));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r11\n",
+          __builtin_offsetof(State, R11));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r12\n",
+          __builtin_offsetof(State, R12));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r13\n",
+          __builtin_offsetof(State, R13));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r14\n",
+          __builtin_offsetof(State, R14));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r15\n",
+          __builtin_offsetof(State, R15));
 
   PrintLoadFlags(out);  // Note: Clobbers RDX.
 
   // XMM registers.
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm0\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm1\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm2\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm3\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm4\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm5\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm6\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm7\n", __builtin_offsetof(State, XMM7));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm8\n", __builtin_offsetof(State, XMM8));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm9\n", __builtin_offsetof(State, XMM9));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm10\n", __builtin_offsetof(State, XMM10));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm11\n", __builtin_offsetof(State, XMM11));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm12\n", __builtin_offsetof(State, XMM12));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm13\n", __builtin_offsetof(State, XMM13));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm14\n", __builtin_offsetof(State, XMM14));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm15\n", __builtin_offsetof(State, XMM15));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm0\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm1\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm2\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm3\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm4\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm5\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm6\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm7\n",
+          __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm8\n",
+          __builtin_offsetof(State, XMM8));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm9\n",
+          __builtin_offsetof(State, XMM9));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm10\n",
+          __builtin_offsetof(State, XMM10));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm11\n",
+          __builtin_offsetof(State, XMM11));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm12\n",
+          __builtin_offsetof(State, XMM12));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm13\n",
+          __builtin_offsetof(State, XMM13));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm14\n",
+          __builtin_offsetof(State, XMM14));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm15\n",
+          __builtin_offsetof(State, XMM15));
 
   // If `RSP` is null then we need to initialize it to our new stack.
-  fprintf(out, "  mov rdx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  mov rdx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSP));
   fprintf(out, "  cmp rdx, 0\n");
   fprintf(out, "  jnz .Lhave_stack\n");
   fprintf(out, "  lea r8, QWORD PTR [__mcsema_stack@TPOFF]\n");
   fprintf(out, "  mov rsi, fs:[0];\n");
-  fprintf(out, "  lea rdx, QWORD PTR [rsi + r8 + %" PRIuMAX "]\n", (kStackSize - 16));
+  fprintf(out, "  lea rdx, QWORD PTR [rsi + r8 + %" PRIuMAX "]\n",
+          (kStackSize - 16));
   fprintf(out, ".Lhave_stack:\n");
 
   // Set up a return address so that when the lifted function returns, it will
@@ -229,11 +286,13 @@ int main(void) {
   fprintf(out, "  pop QWORD PTR [rdx - 16]\n");
 
   // Swap onto the lifted stack. The native `RSP` is now where it should be.
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rsp\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rsp\n",
+          __builtin_offsetof(State, RSP));
   fprintf(out, "  lea rsp, [rdx - 16]\n");
 
   // Set up arg2 as the program counter.
-  fprintf(out, "  mov rsi, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RIP));
+  fprintf(out, "  mov rsi, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RIP));
 
   // Set up arg3 as the memory pointer, which is (for now?) a nullptr.
   fprintf(out, "  xor rdx, rdx\n");
@@ -248,7 +307,8 @@ int main(void) {
   fprintf(out, "  ret\n");
 
   fprintf(out, ".Lfunc_end1:\n");
-  fprintf(out, "  .size __mcsema_attach_call,.Lfunc_end1-__mcsema_attach_call\n");
+  fprintf(out,
+          "  .size __mcsema_attach_call,.Lfunc_end1-__mcsema_attach_call\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -269,46 +329,79 @@ int main(void) {
   // We "undo" that, then swap back to the native stack. When we swap, we
   // save into `State::RSP` where we are in the lifted stack, so that the
   // next attach can continue on where we left off.
-  fprintf(out, "  sub QWORD PTR [rdi + %" PRIuMAX "], 8\n", __builtin_offsetof(State, RSP));
-  fprintf(out, "  xchg [rdi + %" PRIuMAX "], rsp\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  sub QWORD PTR [rdi + %" PRIuMAX "], 8\n",
+          __builtin_offsetof(State, RSP));
+  fprintf(out, "  xchg [rdi + %" PRIuMAX "], rsp\n",
+          __builtin_offsetof(State, RSP));
 
   PrintStoreFlags(out);  // Clobbers RDX.
 
   // General purpose registers.
-  fprintf(out, "  mov rax, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RAX));
-  fprintf(out, "  mov rbx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RBX));
-  fprintf(out, "  mov rcx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RCX));
-  fprintf(out, "  mov rdx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RDX));
-  fprintf(out, "  mov rsi, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RSI));
-  fprintf(out, "  mov rbp, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RBP));
-  fprintf(out, "  mov r8, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R8));
-  fprintf(out, "  mov r9, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R9));
-  fprintf(out, "  mov r10, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R10));
-  fprintf(out, "  mov r11, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R11));
-  fprintf(out, "  mov r12, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R12));
-  fprintf(out, "  mov r13, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R13));
-  fprintf(out, "  mov r14, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R14));
-  fprintf(out, "  mov r15, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R15));
+  fprintf(out, "  mov rax, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RAX));
+  fprintf(out, "  mov rbx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RBX));
+  fprintf(out, "  mov rcx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RCX));
+  fprintf(out, "  mov rdx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RDX));
+  fprintf(out, "  mov rsi, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSI));
+  fprintf(out, "  mov rbp, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RBP));
+  fprintf(out, "  mov r8, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R8));
+  fprintf(out, "  mov r9, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R9));
+  fprintf(out, "  mov r10, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R10));
+  fprintf(out, "  mov r11, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R11));
+  fprintf(out, "  mov r12, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R12));
+  fprintf(out, "  mov r13, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R13));
+  fprintf(out, "  mov r14, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R14));
+  fprintf(out, "  mov r15, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R15));
 
   // XMM registers.
-  fprintf(out, "  movntdqa xmm0, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdqa xmm1, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdqa xmm2, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdqa xmm3, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdqa xmm4, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdqa xmm5, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdqa xmm6, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdqa xmm7, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM7));
-  fprintf(out, "  movntdqa xmm8, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM8));
-  fprintf(out, "  movntdqa xmm9, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM9));
-  fprintf(out, "  movntdqa xmm10, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM10));
-  fprintf(out, "  movntdqa xmm11, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM11));
-  fprintf(out, "  movntdqa xmm12, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM12));
-  fprintf(out, "  movntdqa xmm13, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM13));
-  fprintf(out, "  movntdqa xmm14, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM14));
-  fprintf(out, "  movntdqa xmm15, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM15));
+  fprintf(out, "  movntdqa xmm0, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdqa xmm1, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdqa xmm2, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdqa xmm3, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdqa xmm4, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdqa xmm5, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdqa xmm6, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdqa xmm7, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdqa xmm8, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM8));
+  fprintf(out, "  movntdqa xmm9, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM9));
+  fprintf(out, "  movntdqa xmm10, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM10));
+  fprintf(out, "  movntdqa xmm11, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM11));
+  fprintf(out, "  movntdqa xmm12, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM12));
+  fprintf(out, "  movntdqa xmm13, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM13));
+  fprintf(out, "  movntdqa xmm14, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM14));
+  fprintf(out, "  movntdqa xmm15, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM15));
 
-  fprintf(out, "  mov rdi, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RDI));
+  fprintf(out, "  mov rdi, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RDI));
   fprintf(out, "  ret\n");
 
   fprintf(out, ".Lfunc_end3:\n");
@@ -344,17 +437,18 @@ int main(void) {
   fprintf(out, "  push r14\n");
   fprintf(out, "  push r15\n");
 
-//  fprintf(out, "  push rcx\n");
-//  fprintf(out, "  push rdx\n");
-//  fprintf(out, "  push r8\n");
-//  fprintf(out, "  push r9\n");
-//  fprintf(out, "  push r10\n");
-//  fprintf(out, "  push r11\n");
+  //  fprintf(out, "  push rcx\n");
+  //  fprintf(out, "  push rdx\n");
+  //  fprintf(out, "  push r8\n");
+  //  fprintf(out, "  push r9\n");
+  //  fprintf(out, "  push r10\n");
+  //  fprintf(out, "  push r11\n");
 
 
   // Stash the return address stored on the native stack, the replace it
   // with the re-attach function.
-  fprintf(out, "  mov r15, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  mov r15, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSP));
   fprintf(out, "  push QWORD PTR [r15]\n");
   fprintf(out, "  lea r14, [rip + __mcsema_attach_ret]\n");
   fprintf(out, "  mov QWORD PTR [r15], r14\n");
@@ -368,47 +462,79 @@ int main(void) {
   fprintf(out, "  mov QWORD PTR [r15], rsi\n");
 
   // Swap off-stack, stash the lifted stack pointer.
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rsp\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rsp\n",
+          __builtin_offsetof(State, RSP));
   fprintf(out, "  mov rsp, r15\n");
 
   PrintStoreFlags(out);  // Clobbers RDX.
 
   // (Most) General purpose registers.
-  fprintf(out, "  mov rax, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RAX));
-  fprintf(out, "  mov rbx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RBX));
-  fprintf(out, "  mov rcx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RCX));
-  fprintf(out, "  mov rdx, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RDX));
-  fprintf(out, "  mov rsi, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RSI));
-  fprintf(out, "  mov rbp, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RBP));
-  fprintf(out, "  mov r8, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R8));
-  fprintf(out, "  mov r9, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R9));
-  fprintf(out, "  mov r10, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R10));
-  fprintf(out, "  mov r11, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R11));
-  fprintf(out, "  mov r12, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R12));
-  fprintf(out, "  mov r13, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R13));
-  fprintf(out, "  mov r14, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R14));
-  fprintf(out, "  mov r15, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, R15));
+  fprintf(out, "  mov rax, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RAX));
+  fprintf(out, "  mov rbx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RBX));
+  fprintf(out, "  mov rcx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RCX));
+  fprintf(out, "  mov rdx, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RDX));
+  fprintf(out, "  mov rsi, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSI));
+  fprintf(out, "  mov rbp, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RBP));
+  fprintf(out, "  mov r8, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R8));
+  fprintf(out, "  mov r9, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R9));
+  fprintf(out, "  mov r10, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R10));
+  fprintf(out, "  mov r11, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R11));
+  fprintf(out, "  mov r12, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R12));
+  fprintf(out, "  mov r13, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R13));
+  fprintf(out, "  mov r14, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R14));
+  fprintf(out, "  mov r15, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, R15));
 
   // XMM registers.
-  fprintf(out, "  movntdqa xmm0, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdqa xmm1, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdqa xmm2, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdqa xmm3, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdqa xmm4, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdqa xmm5, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdqa xmm6, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdqa xmm7, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM7));
-  fprintf(out, "  movntdqa xmm8, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM8));
-  fprintf(out, "  movntdqa xmm9, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM9));
-  fprintf(out, "  movntdqa xmm10, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM10));
-  fprintf(out, "  movntdqa xmm11, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM11));
-  fprintf(out, "  movntdqa xmm12, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM12));
-  fprintf(out, "  movntdqa xmm13, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM13));
-  fprintf(out, "  movntdqa xmm14, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM14));
-  fprintf(out, "  movntdqa xmm15, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, XMM15));
+  fprintf(out, "  movntdqa xmm0, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdqa xmm1, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdqa xmm2, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdqa xmm3, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdqa xmm4, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdqa xmm5, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdqa xmm6, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdqa xmm7, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdqa xmm8, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM8));
+  fprintf(out, "  movntdqa xmm9, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM9));
+  fprintf(out, "  movntdqa xmm10, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM10));
+  fprintf(out, "  movntdqa xmm11, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM11));
+  fprintf(out, "  movntdqa xmm12, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM12));
+  fprintf(out, "  movntdqa xmm13, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM13));
+  fprintf(out, "  movntdqa xmm14, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM14));
+  fprintf(out, "  movntdqa xmm15, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, XMM15));
 
   // Swap out RDI.
-  fprintf(out, "  mov rdi, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RDI));
+  fprintf(out, "  mov rdi, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RDI));
 
   // Code above put the native target address (stored in RDI on entry to
   // `__remill_function_call`) on the stack, just below the return address,
@@ -418,7 +544,9 @@ int main(void) {
   fprintf(out, "  ret\n");
 
   fprintf(out, ".Lfunc_end5:\n");
-  fprintf(out, "  .size __remill_function_call,.Lfunc_end5-__remill_function_call\n");
+  fprintf(
+      out,
+      "  .size __remill_function_call,.Lfunc_end5-__remill_function_call\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -432,49 +560,82 @@ int main(void) {
   // Copy RSI, then store the address of the reg state struct into RSI for
   // easier indexing later on. Also set up the `FS` segment register so that
   // TLS works :-)
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n", __builtin_offsetof(State, RDI));
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n", __builtin_offsetof(State, RSI));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n",
+          __builtin_offsetof(State, RDI));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n",
+          __builtin_offsetof(State, RSI));
 
   fprintf(out, "  mov rdi, QWORD PTR fs:[0]\n");
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n", __builtin_offsetof(State, FS_BASE));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n",
+          __builtin_offsetof(State, FS_BASE));
   fprintf(out, "  lea rsi, [__mcsema_reg_state@TPOFF]\n");
   fprintf(out, "  lea rdi, QWORD PTR [rsi + rdi]\n");
 
   // General purpose registers.
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n", __builtin_offsetof(State, RAX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbx\n", __builtin_offsetof(State, RBX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rcx\n", __builtin_offsetof(State, RCX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rdx\n", __builtin_offsetof(State, RDX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbp\n", __builtin_offsetof(State, RBP));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r8\n", __builtin_offsetof(State, R8));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r9\n", __builtin_offsetof(State, R9));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r10\n", __builtin_offsetof(State, R10));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r11\n", __builtin_offsetof(State, R11));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r12\n", __builtin_offsetof(State, R12));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r13\n", __builtin_offsetof(State, R13));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r14\n", __builtin_offsetof(State, R14));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r15\n", __builtin_offsetof(State, R15));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n",
+          __builtin_offsetof(State, RAX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbx\n",
+          __builtin_offsetof(State, RBX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rcx\n",
+          __builtin_offsetof(State, RCX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rdx\n",
+          __builtin_offsetof(State, RDX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbp\n",
+          __builtin_offsetof(State, RBP));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r8\n",
+          __builtin_offsetof(State, R8));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r9\n",
+          __builtin_offsetof(State, R9));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r10\n",
+          __builtin_offsetof(State, R10));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r11\n",
+          __builtin_offsetof(State, R11));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r12\n",
+          __builtin_offsetof(State, R12));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r13\n",
+          __builtin_offsetof(State, R13));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r14\n",
+          __builtin_offsetof(State, R14));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r15\n",
+          __builtin_offsetof(State, R15));
 
   // Swap into the mcsema stack.
-  fprintf(out, "  xchg rsp, [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  xchg rsp, [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSP));
 
   // XMM registers.
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm0\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm1\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm2\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm3\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm4\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm5\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm6\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm7\n", __builtin_offsetof(State, XMM7));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm8\n", __builtin_offsetof(State, XMM8));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm9\n", __builtin_offsetof(State, XMM9));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm10\n", __builtin_offsetof(State, XMM10));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm11\n", __builtin_offsetof(State, XMM11));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm12\n", __builtin_offsetof(State, XMM12));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm13\n", __builtin_offsetof(State, XMM13));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm14\n", __builtin_offsetof(State, XMM14));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm15\n", __builtin_offsetof(State, XMM15));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm0\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm1\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm2\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm3\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm4\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm5\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm6\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm7\n",
+          __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm8\n",
+          __builtin_offsetof(State, XMM8));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm9\n",
+          __builtin_offsetof(State, XMM9));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm10\n",
+          __builtin_offsetof(State, XMM10));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm11\n",
+          __builtin_offsetof(State, XMM11));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm12\n",
+          __builtin_offsetof(State, XMM12));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm13\n",
+          __builtin_offsetof(State, XMM13));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm14\n",
+          __builtin_offsetof(State, XMM14));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm15\n",
+          __builtin_offsetof(State, XMM15));
 
   PrintLoadFlags(out);  // Note: Clobbers RDX.
 
@@ -488,15 +649,16 @@ int main(void) {
   //    48    stashed rbx
 
   // Restore emulated return address.
-  fprintf(out, "  pop QWORD PTR [rdi + %" PRIuMAX "]\n", __builtin_offsetof(State, RIP));
+  fprintf(out, "  pop QWORD PTR [rdi + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RIP));
 
-//
-//  fprintf(out, "  pop r11\n");
-//  fprintf(out, "  pop r10\n");
-//  fprintf(out, "  pop r9\n");
-//  fprintf(out, "  pop r8\n");
-//  fprintf(out, "  pop rdx\n");
-//  fprintf(out, "  pop rcx\n");
+  //
+  //  fprintf(out, "  pop r11\n");
+  //  fprintf(out, "  pop r10\n");
+  //  fprintf(out, "  pop r9\n");
+  //  fprintf(out, "  pop r8\n");
+  //  fprintf(out, "  pop rdx\n");
+  //  fprintf(out, "  pop rcx\n");
 
 
   // Callee-saved registers.
@@ -529,59 +691,96 @@ int main(void) {
   fprintf(out, ".Lfunc_begin10:\n");
   fprintf(out, ".cfi_startproc\n");
 
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n", __builtin_offsetof(State, RDI));
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n", __builtin_offsetof(State, RSI));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n",
+          __builtin_offsetof(State, RDI));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rsi\n",
+          __builtin_offsetof(State, RSI));
 
   fprintf(out, "  mov rdi, QWORD PTR fs:[0]\n");
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n", __builtin_offsetof(State, FS_BASE));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rdi\n",
+          __builtin_offsetof(State, FS_BASE));
   fprintf(out, "  lea rsi, [__mcsema_reg_state@TPOFF]\n");
   fprintf(out, "  lea rdi, QWORD PTR [rsi + rdi]\n");
 
   // General purpose registers.
   //fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n", __builtin_offsetof(State, RAX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbx\n", __builtin_offsetof(State, RBX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rcx\n", __builtin_offsetof(State, RCX));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rdx\n", __builtin_offsetof(State, RDX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rbx\n",
+          __builtin_offsetof(State, RBX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rcx\n",
+          __builtin_offsetof(State, RCX));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rdx\n",
+          __builtin_offsetof(State, RDX));
 
   // Sets the native stack and base pointers
-  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n", __builtin_offsetof(State, RDI));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n", __builtin_offsetof(State, RSP));
-  fprintf(out, "  add QWORD PTR [rdi + %" PRIuMAX "], 8\n", __builtin_offsetof(State, RSP));
-  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n", __builtin_offsetof(State, RSI));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n", __builtin_offsetof(State, RBP));
+  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RDI));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n",
+          __builtin_offsetof(State, RSP));
+  fprintf(out, "  add QWORD PTR [rdi + %" PRIuMAX "], 8\n",
+          __builtin_offsetof(State, RSP));
+  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSI));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], rax\n",
+          __builtin_offsetof(State, RBP));
 
 
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r8\n", __builtin_offsetof(State, R8));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r9\n", __builtin_offsetof(State, R9));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r10\n", __builtin_offsetof(State, R10));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r11\n", __builtin_offsetof(State, R11));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r12\n", __builtin_offsetof(State, R12));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r13\n", __builtin_offsetof(State, R13));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r14\n", __builtin_offsetof(State, R14));
-  fprintf(out, "  mov [rdi + %" PRIuMAX "], r15\n", __builtin_offsetof(State, R15));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r8\n",
+          __builtin_offsetof(State, R8));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r9\n",
+          __builtin_offsetof(State, R9));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r10\n",
+          __builtin_offsetof(State, R10));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r11\n",
+          __builtin_offsetof(State, R11));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r12\n",
+          __builtin_offsetof(State, R12));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r13\n",
+          __builtin_offsetof(State, R13));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r14\n",
+          __builtin_offsetof(State, R14));
+  fprintf(out, "  mov [rdi + %" PRIuMAX "], r15\n",
+          __builtin_offsetof(State, R15));
 
   // XMM registers.
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm0\n", __builtin_offsetof(State, XMM0));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm1\n", __builtin_offsetof(State, XMM1));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm2\n", __builtin_offsetof(State, XMM2));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm3\n", __builtin_offsetof(State, XMM3));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm4\n", __builtin_offsetof(State, XMM4));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm5\n", __builtin_offsetof(State, XMM5));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm6\n", __builtin_offsetof(State, XMM6));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm7\n", __builtin_offsetof(State, XMM7));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm8\n", __builtin_offsetof(State, XMM8));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm9\n", __builtin_offsetof(State, XMM9));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm10\n", __builtin_offsetof(State, XMM10));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm11\n", __builtin_offsetof(State, XMM11));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm12\n", __builtin_offsetof(State, XMM12));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm13\n", __builtin_offsetof(State, XMM13));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm14\n", __builtin_offsetof(State, XMM14));
-  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm15\n", __builtin_offsetof(State, XMM15));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm0\n",
+          __builtin_offsetof(State, XMM0));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm1\n",
+          __builtin_offsetof(State, XMM1));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm2\n",
+          __builtin_offsetof(State, XMM2));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm3\n",
+          __builtin_offsetof(State, XMM3));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm4\n",
+          __builtin_offsetof(State, XMM4));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm5\n",
+          __builtin_offsetof(State, XMM5));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm6\n",
+          __builtin_offsetof(State, XMM6));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm7\n",
+          __builtin_offsetof(State, XMM7));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm8\n",
+          __builtin_offsetof(State, XMM8));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm9\n",
+          __builtin_offsetof(State, XMM9));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm10\n",
+          __builtin_offsetof(State, XMM10));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm11\n",
+          __builtin_offsetof(State, XMM11));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm12\n",
+          __builtin_offsetof(State, XMM12));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm13\n",
+          __builtin_offsetof(State, XMM13));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm14\n",
+          __builtin_offsetof(State, XMM14));
+  fprintf(out, "  movntdq [rdi + %" PRIuMAX "], xmm15\n",
+          __builtin_offsetof(State, XMM15));
 
   fprintf(out, "  ret\n");
   fprintf(out, "  ud2\n");
   fprintf(out, ".Lfunc_end10:\n");
-  fprintf(out, "  .size __mcsema_exception_ret,.Lfunc_end10-__mcsema_exception_ret\n");
+  fprintf(
+      out,
+      "  .size __mcsema_exception_ret,.Lfunc_end10-__mcsema_exception_ret\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -590,10 +789,13 @@ int main(void) {
   fprintf(out, "  .type __mcsema_get_stack_pointer,@function\n");
   fprintf(out, "__mcsema_get_stack_pointer:\n");
   fprintf(out, "  .cfi_startproc\n");
-  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n", __builtin_offsetof(State, RSP));
+  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RSP));
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end20:\n");
-  fprintf(out, "  .size __mcsema_get_stack_pointer,.Lfunc_end20-__mcsema_get_stack_pointer\n");
+  fprintf(
+      out,
+      "  .size __mcsema_get_stack_pointer,.Lfunc_end20-__mcsema_get_stack_pointer\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -602,10 +804,13 @@ int main(void) {
   fprintf(out, "  .type __mcsema_get_frame_pointer,@function\n");
   fprintf(out, "__mcsema_get_frame_pointer:\n");
   fprintf(out, "  .cfi_startproc\n");
-  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n", __builtin_offsetof(State, RBP));
+  fprintf(out, "  mov rax, fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "]\n",
+          __builtin_offsetof(State, RBP));
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end21:\n");
-  fprintf(out, "  .size __mcsema_get_frame_pointer,.Lfunc_end21-__mcsema_get_frame_pointer\n");
+  fprintf(
+      out,
+      "  .size __mcsema_get_frame_pointer,.Lfunc_end21-__mcsema_get_frame_pointer\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -614,11 +819,14 @@ int main(void) {
   fprintf(out, "  .type __mcsema_get_type_index,@function\n");
   fprintf(out, "__mcsema_get_type_index:\n");
   fprintf(out, "  .cfi_startproc\n");
-  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rax\n", __builtin_offsetof(State, RAX));
+  fprintf(out, "  mov fs:[__mcsema_reg_state@TPOFF + %" PRIuMAX "], rax\n",
+          __builtin_offsetof(State, RAX));
   fprintf(out, "  mov rax, rdx\n");
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end22:\n");
-  fprintf(out, "  .size __mcsema_get_type_index,.Lfunc_end22-__mcsema_get_type_index\n");
+  fprintf(
+      out,
+      "  .size __mcsema_get_type_index,.Lfunc_end22-__mcsema_get_type_index\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -633,7 +841,9 @@ int main(void) {
   fprintf(out, "  lea rax, [rax + rdx]\n");
   fprintf(out, "  ret\n");
   fprintf(out, ".Lfunc_end6:\n");
-  fprintf(out, "  .size __mcsema_debug_get_reg_state,.Lfunc_end6-__mcsema_debug_get_reg_state\n");
+  fprintf(
+      out,
+      "  .size __mcsema_debug_get_reg_state,.Lfunc_end6-__mcsema_debug_get_reg_state\n");
   fprintf(out, "  .cfi_endproc\n");
   fprintf(out, "\n");
 
@@ -664,5 +874,3 @@ int main(void) {
 
 //  // Restore stack alignment
 //  fprintf(out, "  pop rsp\n");
-
-
