@@ -1,21 +1,21 @@
 /*
- * Copyright (c) 2017 Trail of Bits, Inc.
+ * Copyright (c) 2020 Trail of Bits, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef MCSEMA_BC_CALLBACK_H_
-#define MCSEMA_BC_CALLBACK_H_
+#pragma once
 
 namespace llvm {
 
@@ -23,29 +23,16 @@ class Function;
 
 }  // namespace llvm
 namespace mcsema {
-struct NativeObject;
-
-// Get a callback function for an internal function that can be referenced by
-// internal code.
-llvm::Function *GetNativeToLiftedCallback(const NativeObject *cfg_func);
-
-// Get a callback function for an internal function that can be referenced by
-// external code.
-llvm::Function *GetNativeToLiftedEntryPoint(const NativeObject *cfg_func);
+struct NativeFunction;
 
 // Get a callback function for an external function that can be referenced by
 // internal code.
-llvm::Function *GetLiftedToNativeExitPoint(const NativeObject *cfg_func);
+llvm::Function *GetLiftedToNativeExitPoint(const NativeFunction *cfg_func);
 
-enum ExitPointKind {
-  kExitPointJump,
-  kExitPointFunctionCall
-};
+enum ExitPointKind { kExitPointJump, kExitPointFunctionCall };
 
 // Get a function that goes from the current lifted state into native state,
 // where we don't know where the native destination actually is.
 llvm::Function *GetLiftedToNativeExitPoint(ExitPointKind);
 
 }  // namespace mcsema
-
-#endif  // MCSEMA_BC_CALLBACK_H_
