@@ -1,37 +1,33 @@
 /*
- * Copyright (c) 2017 Trail of Bits, Inc.
+ * Copyright (c) 2020 Trail of Bits, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef MCSEMA_BC_EXTERNAL_H_
-#define MCSEMA_BC_EXTERNAL_H_
+#pragma once
 
 #include <llvm/IR/CallingConv.h>
 
 namespace llvm {
 namespace CallingConv {
-  enum {
-    McSemaCall = 144
-  };
-} // namespace CallingConv
-} // namespace llvm
+enum { McSemaCall = 144 };
+}  // namespace CallingConv
+}  // namespace llvm
 
 namespace mcsema {
 struct NativeModule;
 
-void DeclareExternals(const NativeModule *cfg_module);
+//void DeclareExternals(const NativeModule *cfg_module);
 
 }  // namespace mcsema
-
-#endif  // MCSEMA_BC_EXTERNAL_H_
