@@ -1,35 +1,44 @@
 /*
- * Copyright (c) 2017 Trail of Bits, Inc.
+ * Copyright (c) 2020 Trail of Bits, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include "Legacy.h"
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wsign-conversion"
+#pragma clang diagnostic ignored "-Wconversion"
+#pragma clang diagnostic ignored "-Wold-style-cast"
+#pragma clang diagnostic ignored "-Wdocumentation"
+#pragma clang diagnostic ignored "-Wswitch-enum"
 #include <gflags/gflags.h>
 #include <glog/logging.h>
-
-#include <vector>
-
 #include <llvm/IR/Constants.h>
-#include <llvm/IR/Instructions.h>
 #include <llvm/IR/IRBuilder.h>
+#include <llvm/IR/Instructions.h>
 #include <llvm/IR/LLVMContext.h>
 #include <llvm/IR/Module.h>
 #include <llvm/IR/Type.h>
+#pragma clang diagnostic pop
 
-#include "remill/Arch/Arch.h"
-#include "remill/BC/ABI.h"
-#include "remill/BC/Util.h"
-#include "remill/BC/Version.h"
+#include <remill/Arch/Arch.h>
+#include <remill/BC/ABI.h>
+#include <remill/BC/Util.h>
+#include <remill/BC/Version.h>
+
+#include <vector>
 
 #include "mcsema/Arch/Arch.h"
 #include "mcsema/BC/Callback.h"
@@ -44,24 +53,6 @@ DEFINE_string(pc_annotation, "",
 namespace mcsema {
 namespace legacy {
 namespace {
-
-// Remove calls to error-related intrinsics.
-static void ImplementErrorIntrinsic(const char *name) {
-  auto func = gModule->getFunction(name);
-  if (!func) {
-    return;
-  }
-
-  auto void_type = llvm::Type::getVoidTy(*gContext);
-  auto abort_func = gModule->getOrInsertFunction(
-      "abort", llvm::FunctionType::get(void_type, false));
-
-  func->setLinkage(llvm::GlobalValue::InternalLinkage);
-
-  llvm::IRBuilder<> ir(llvm::BasicBlock::Create(*gContext, "", func));
-  ir.CreateCall(abort_func);
-  ir.CreateRet(remill::NthArgument(func, remill::kMemoryPointerArgNum));
-}
 
 // Create the node for a `mcsema_real_eip` annotation.
 static llvm::MDNode *CreateInstAnnotation(llvm::Function *F, uint64_t addr) {
@@ -144,11 +135,6 @@ void PropagateInstAnnotations(void) {
       }
     }
   }
-}
-
-void DowngradeModule(void) {
-  ImplementErrorIntrinsic("__remill_error");
-  ImplementErrorIntrinsic("__remill_missing_block");
 }
 
 }  // namespace legacy
