@@ -1,27 +1,27 @@
 /*
- * Copyright (c) 2017 Trail of Bits, Inc.
+ * Copyright (c) 2020 Trail of Bits, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef MCSEMA_BC_UTIL_H_
-#define MCSEMA_BC_UTIL_H_
+#pragma once
+
+#include <llvm/IR/IRBuilder.h>
 
 #include <cstdint>
 #include <list>
 #include <vector>
-
-#include <llvm/IR/IRBuilder.h>
 
 #include "mcsema/CFG/CFG.h"
 
@@ -40,20 +40,25 @@ namespace mcsema {
 
 struct NativeSegment;
 
-extern llvm::LLVMContext *gContext;
+extern std::shared_ptr<llvm::LLVMContext> gContext;
 extern llvm::IntegerType *gWordType;
-extern llvm::Module *gModule;
-
+extern std::unique_ptr<llvm::Module> gModule;
+extern llvm::Constant *gZero;
+extern uint64_t gWordMask;
 
 llvm::Value *GetConstantInt(unsigned size, uint64_t value);
 
-// Return the type of a lifted function.
-llvm::FunctionType *LiftedFunctionType(void);
+// Get a lifted representation of a reference (in code) to `ea`.
+llvm::Constant *LiftXrefInCode(uint64_t ea);
 
-// Translate `ea` into an LLVM value that is an address that points into the
-// lifted segment associated with `seg`.
-llvm::Constant *LiftEA(const NativeSegment *seg, uint64_t ea);
+// Get a lifted representation of a reference (in data) to `ea`.
+llvm::Constant *LiftXrefInData(const NativeSegment *cfg_seg, uint64_t ea,
+                               bool cast_to_int = true);
+
+// Create a global register state pointer to pass to lifted functions.
+llvm::Constant *GetStatePointer(void);
+
+// Return the address of the base of the TLS data.
+llvm::Value *GetTLSBaseAddress(llvm::IRBuilder<> &ir);
 
 }  // namespace mcsema
-
-#endif  // MCSEMA_BC_UTIL_H_
