@@ -1,17 +1,18 @@
 #!/usr/bin/env python
-# Copyright (c) 2017 Trail of Bits, Inc.
+# Copyright (c) 2020 Trail of Bits, Inc.
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
 #
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import argparse
 import hashlib
@@ -28,7 +29,7 @@ except:
 
 def binary_info(binary):
   res = subprocess.check_output(['file', binary])
-  is_pie = 'LSB shared object' in res or 'Mach-O 64' in res
+  is_pie = 'LSB shared object' in res or 'Mach-O 64' in res or 'LSB pie executable' in res
   address_size = 64
 
   if 'aarch64' in res:
@@ -55,7 +56,7 @@ def binary_libraries(binary):
     path_and_addr = path_and_addr.strip(" ")
     if not path_and_addr.startswith("/"):
       continue
-    
+
     lib = " ".join(path_and_addr.split(" ")[:-1])
     lib = os.path.realpath(lib)
 
@@ -139,14 +140,14 @@ def main():
   bc_dir = os.path.join(args.workspace_dir, 'bc')
   log_dir = os.path.join(args.workspace_dir, 'log')
 
-  print "mkdir -p {}".format(args.workspace_dir)
-  print "mkdir -p {}".format(bin_dir)
-  print "mkdir -p {}".format(lifted_obj_dir)
-  print "mkdir -p {}".format(lib_dir)
-  print "mkdir -p {}".format(obj_dir)
-  print "mkdir -p {}".format(cfg_dir)
-  print "mkdir -p {}".format(bc_dir)
-  print "mkdir -p {}".format(log_dir)
+  print("mkdir -p {}".format(args.workspace_dir))
+  print("mkdir -p {}".format(bin_dir))
+  print("mkdir -p {}".format(lifted_obj_dir))
+  print("mkdir -p {}".format(lib_dir))
+  print("mkdir -p {}".format(obj_dir))
+  print("mkdir -p {}".format(cfg_dir))
+  print("mkdir -p {}".format(bc_dir))
+  print("mkdir -p {}".format(log_dir))
 
   make_directory(args.workspace_dir)
   make_directory(bin_dir)
@@ -168,8 +169,8 @@ def main():
   log = os.path.join(log_dir, "{}.log".format(path_hash))
 
   # Copy the binary into the workspace's object directory.
-  print "cp {} {}".format(path, binary)
-  print "chmod a+x {}".format(binary)
+  print("cp {} {}".format(path, binary))
+  print("chmod a+x {}".format(binary))
   if not os.path.isfile(binary):
     shutil.copyfile(args.binary, binary)
     make_executable(binary)
@@ -195,10 +196,10 @@ def main():
     except:
       pass
 
-    print "cp {} {}".format(path, library)
-    print "chmod a+x {}".format(library)
-    print "rm {}".format(sym_name)
-    print "ln {} {}".format(library, sym_name)
+    print("cp {} {}".format(path, library))
+    print("chmod a+x {}".format(library))
+    print("rm {}".format(sym_name))
+    print("ln {} {}".format(library, sym_name))
 
     libs.append(sym_name)
 
@@ -234,7 +235,7 @@ def main():
 
   disass_args.extend(command_args)
 
-  print " ".join(disass_args)
+  print(" ".join(disass_args))
   ret = subprocess.call(disass_args)
   if ret:
     return ret
@@ -254,7 +255,7 @@ def main():
   if args.legacy_mode:
     mcsema_lift_args.append('--legacy_mode')
 
-  print " ".join(mcsema_lift_args)
+  print(" ".join(mcsema_lift_args))
   ret = subprocess.call(mcsema_lift_args)
   if ret:
     return ret
@@ -285,7 +286,7 @@ def main():
     clang_args.append(lib)
 
   # Compile back to an executable.
-  print " ".join(clang_args)
+  print(" ".join(clang_args))
   ret = subprocess.call(clang_args)
   if ret:
     return ret
