@@ -1,21 +1,26 @@
-# Copyright (c) 2017 Trail of Bits, Inc.
+# Copyright (c) 2020 Trail of Bits, Inc.
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
 #
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from refs import *
 
+import struct
+
 _FIRST_JUMP_TABLE_ENTRY = {}
 _JUMP_TABLE_ENTRY = {}
+_IS_JUMP_TABLE_ENTRY = set()
+_FUDGE_FACTOR = 256
 
 class JumpTable(object):
   """Represents generic info known about a particular jump table."""
@@ -39,9 +44,15 @@ class JumpTable(object):
       idc.add_cref(self.inst_ea, target_ea, idc.XREF_USER | idc.fl_JN)
       max_ea = max(max_ea, entry_ea + self.entry_size)
     
+    entry_type = ida_bytes.FF_QWORD
+    if self.entry_size == 4:
+      entry_type = ida_bytes.FF_DWORD
+
     # Make sure each entry is seen as referenced.
     for entry_ea, target_ea in entries.items():
-      idc.add_dref(self.inst_ea, entry_ea, idc.dr_I)
+      idc.create_data(entry_ea, entry_type, self.entry_size, idaapi.BADADDR)
+      idc.add_dref(entry_ea, target_ea, idc.XREF_USER|idc.dr_I)
+      idc.add_dref(self.inst_ea, entry_ea, idc.XREF_USER|idc.dr_I)
 
     # Make sure that the contents of the table are no longer considered
     # code (only affects `is_code`). This is only meaningful if the table
@@ -49,9 +60,12 @@ class JumpTable(object):
     for ea_into_table in xrange(self.table_ea, max_ea):
       mark_as_not_code(ea_into_table)
       _FIRST_JUMP_TABLE_ENTRY[ea_into_table] = self.table_ea
+      _IS_JUMP_TABLE_ENTRY.add(ea_into_table)
 
     _JUMP_TABLE_ENTRY[self.table_ea] = self
     idaapi.auto_wait()
+
+_IS_TARGETED_BY_JUMP_TABLE = set()
 
 class JumpTableBuilder(object):
   _READERS = {
@@ -60,6 +74,8 @@ class JumpTableBuilder(object):
   }
 
   def __init__(self, inst, binary_is_pie):
+    import util
+
     self.entry_size = 0
     self.entry_mult = 1
     self.table_ea = idc.BADADDR
@@ -72,13 +88,24 @@ class JumpTableBuilder(object):
 
   def read_entry(self, entry_ea):
     data = (self._READERS[self.entry_size])(entry_ea)
-    data &= ((1 << (self.entry_size * 8)) - 1)
+    data &= ((1 << int(self.entry_size * 8)) - 1)
     data += self.offset * self.offset_mult
     data &= 0xFFFFFFFFFFFFFFFF
     if 4 == self.entry_size and (self.offset & 0xFFFFFFFF) == self.offset:
       data &= 0xFFFFFFFF
-
     return data
+
+def _check_entry_target_ea(target_ea, min_ea, max_ea):
+  global _FUDGE_FACTOR
+  min_ea = max(0, min_ea - _FUDGE_FACTOR)
+  max_ea += _FUDGE_FACTOR
+  if min_ea > target_ea or max_ea <= target_ea:
+    return False
+
+  if not is_block_or_instruction_head(target_ea):
+    return False
+
+  return True
 
 def get_default_jump_table_entries(builder):
   """Return the 'default' jump table entries, based on IDA's ability to
@@ -98,6 +125,8 @@ def get_default_jump_table_entries(builder):
 def get_num_jump_table_entries(builder):
   """Try to get the number of entries in a jump table. This will use some
   base set of entries."""
+  global _IS_TARGETED_BY_JUMP_TABLE
+
   curr_num_targets = len(builder.candidate_target_eas)
 
   DEBUG("Checking if jump table at {:x} has more than {} entries".format(
@@ -121,6 +150,8 @@ def get_num_jump_table_entries(builder):
       DEBUG("  ERROR jump table {:x} entry candidate {} target {:x} is not sane!".format(
           builder.table_ea, i, curr_target))
       continue
+    
+    _IS_TARGETED_BY_JUMP_TABLE.add(curr_target)
 
     targ_min_ea, targ_max_ea = get_function_bounds(curr_target)
     if targ_min_ea >= max_ea or targ_max_ea <= min_ea:
@@ -145,9 +176,15 @@ def get_num_jump_table_entries(builder):
 
   # The candidate targets have given us some rough bounds on the function,
   # now lets go and check all the targets
-  max_i = max(curr_num_targets, 1024)
+  max_i = max(curr_num_targets, 2048)
   entry_addr = builder.table_ea
   table_seg_ea = idc.get_segm_start(builder.table_ea)
+  stop_at = max_i
+
+  entry_type = ida_bytes.FF_QWORD
+  if builder.entry_size == 4:
+    entry_type = ida_bytes.FF_DWORD
+
   for i in xrange(max_i):
 
     # Make sure we don't read across a segment (e.g. if the jump table is the
@@ -157,6 +194,9 @@ def get_num_jump_table_entries(builder):
         break
     except:
       break
+
+    #if entry_addr in _FIRST_JUMP_TABLE_ENTRY:
+    #  DEBUG("  Entry at {:x} is already part of another jump table".format(entry_addr))
 
     entry_data = builder.read_entry(entry_addr)
     next_entry_addr = entry_addr + (builder.entry_size * builder.entry_mult)
@@ -174,9 +214,18 @@ def get_num_jump_table_entries(builder):
       DEBUG("    Not an entry, the target {:x} isn't sane.".format(entry_data))
       break
     
-    elif min_ea > entry_data or entry_data >= max_ea:
-      DEBUG("    Not an entry, the target {:x} is out of range.".format(entry_data))
-      break
+    elif not _check_entry_target_ea(entry_data, min_ea, max_ea):
+      DEBUG("    Not an entry, the target {:x} is out of range [{:x}, {:x})".format(
+          entry_data, min_ea, max_ea))
+      stop_at = min(stop_at, i)
+
+      if entry_data != get_reference_target(entry_addr):
+        idc.create_data(entry_addr, entry_type, builder.entry_size, idaapi.BADADDR)
+        idc.add_dref(entry_addr, entry_data, idc.XREF_USER|idc.dr_I)
+
+      _IS_TARGETED_BY_JUMP_TABLE.add(entry_data)
+      for entry_addr_sub_ea in xrange(entry_addr, entry_addr + builder.entry_size):
+        _IS_JUMP_TABLE_ENTRY.add(entry_addr_sub_ea)
 
     # We will assume that any reference to the data in here means
     # that we've gone and found the end of a table.
@@ -185,8 +234,9 @@ def get_num_jump_table_entries(builder):
     #      a reference into the Nth byte of what could be the
     #      next address.
     elif len(list(idautils.DataRefsTo(entry_addr))):
-      DEBUG("    Ignoring entry {:x} is referenced by data.".format(entry_data))
-      break
+      if 0 < i:  # The first entry's reference might be a dref in an instruction.
+        DEBUG("    Ignoring entry {:x} is referenced by data.".format(entry_data))
+        break
     elif len(list(idautils.CodeRefsTo(entry_addr, 0))):
       DEBUG("    Ignoring entry {:x} is referenced by code (0).".format(entry_data))
       break
@@ -194,11 +244,18 @@ def get_num_jump_table_entries(builder):
       DEBUG("    Ignoring entry {:x} is referenced by code (1).".format(entry_data))
       break
 
+    # Widen the bounds if the fudge factor came into play.
+    min_ea = min(min_ea, entry_data)
+    max_ea = max(max_ea, entry_data)
+    if stop_at < i < max_i:
+      stop_at = max_i
+
     entry_addr = next_entry_addr
 
+  i = min(i, stop_at)
   if i != curr_num_targets:
-    DEBUG("Jump table at {:x} actually has {} entries".format(
-        builder.table_ea, i))
+    DEBUG("Jump table at {:x} actually has {} != {} entries".format(
+        builder.table_ea, i, curr_num_targets))
 
   return i
 
@@ -230,21 +287,16 @@ def try_get_simple_jump_table_reader(builder):
 
           target_ea = builder.read_entry(builder.table_ea)
 
-          if min_ea > target_ea or max_ea <= target_ea:
-            continue
-
-          if not is_block_or_instruction_head(target_ea):
-            continue
-
           # Read the second entry from the table as a way of verifying our
           # guesses on things like the offset multiplier and entry multiplier.
           next_entry_addr = builder.table_ea + \
-                            (builder.entry_size * builder.entry_mult)
+                            int(builder.entry_size * builder.entry_mult)
 
-          target_ea = builder.read_entry(next_entry_addr)
-          if min_ea <= target_ea < max_ea and \
-             is_block_or_instruction_head(target_ea):
-             return True
+          if _check_entry_target_ea(target_ea, min_ea, max_ea):
+            target_ea = builder.read_entry(next_entry_addr)
+            # Additional check is added to avoid target_ea becoming valid if the data at next_entry_addr is 0
+            if (target_ea != builder.offset)  and _check_entry_target_ea(target_ea, min_ea, max_ea):
+              return True
 
   return False
 
@@ -285,7 +337,11 @@ def get_ida_jump_table_reader(builder, si):
     builder.entry_mult = -1
 
   DEBUG("IDA inferred jump table entry size: {}".format(builder.entry_size))
-
+  
+  if builder.entry_size not in (4, 8):
+    builder.entry_size = get_address_size_in_bits() // 8
+    DEBUG("Using jump table entry size {} instead".format(builder.entry_size))
+    
   # Check if this is an offset based jump table, and if so, create an
   # appropriate wrapper that uses the displacement from the table base
   # address to find the actual jump target.
@@ -293,7 +349,8 @@ def get_ida_jump_table_reader(builder, si):
     builder.offset = si.elbase
     
     # Figure out if we need to subtract the offset instead of add it.
-    if (si.flags2 & idaapi.SWI2_SUBTRACT) == idaapi.SWI2_SUBTRACT:
+    SWI2_SUBTRACT = idaapi.SWI_SUBTRACT >> 16
+    if (si.flags & SWI2_SUBTRACT) == SWI2_SUBTRACT:
       builder.offset_mult = -1
 
     DEBUG("IDA inferred jump table offset: {:x}".format(builder.offset))
@@ -327,35 +384,62 @@ def get_manual_jump_table_reader(builder):
   even if it's not explicitly referenced in the current instruction.
   This handles the case where we see something like a `mov` or an `lea`
   of the table base address that happens before the actual `jmp`."""
+  if not is_invalid_ea(builder.table_ea):
+    if try_get_simple_jump_table_reader(builder):
+      return True
+
+  inst_ea = builder.jump_ea
+  block_eas = list()
+  block_eas.append(inst_ea)
+  for i in xrange(8):
+    prev_head_ea = idc.prev_head(inst_ea)
+
+    for xref_ea in crefs_to(inst_ea):
+      if prev_head_ea != xref_ea:
+        inst, _ = decode_instruction(xref_ea)
+        if has_delayed_slot(inst):
+          block_eas.append(inst.ea + inst.size)
+        else:
+          block_eas.append(xref_ea)
+
+    inst_ea = prev_head_ea
+
   ret = False
   next_inst_ea = builder.jump_ea
   found_ref_eas = set()
-  for i in xrange(5):
-    inst_ea = next_inst_ea
-    next_inst_ea = idc.prev_head(inst_ea)
-    if inst_ea == idc.BADADDR:
-      break
 
-    elif builder.jump_ea != inst_ea and is_control_flow(inst_ea):
-      break
+  for block_ea in block_eas:
+    next_inst_ea = block_ea
+    for i in xrange(10):
+      inst_ea = next_inst_ea
+      next_inst_ea = idc.prev_head(inst_ea)
+      if inst_ea == idc.BADADDR:
+        break
 
-    refs = get_instruction_references(inst_ea, builder.binary_is_pie)
-    if not len(refs):
-      continue
+      elif is_noreturn_external_function(inst_ea):
+        break
 
-    found_ref_eas.add((inst_ea, refs[0].ea))
+      elif builder.jump_ea != inst_ea and is_control_flow(inst_ea):
+        continue
 
-    builder.table_ea = refs[0].ea
-    builder.offset = 0
+      refs = get_instruction_references(inst_ea, builder.binary_is_pie)
+      if not len(refs):
+        continue
 
-    # Don't treat things like thunks to be tables.
-    if is_thunk(builder.table_ea) or is_external_segment(builder.table_ea):
-      builder.table_ea = idc.BADADDR
-      continue
+      found_ref_eas.add((inst_ea, refs[0].ea))
 
-    if try_get_simple_jump_table_reader(builder):
-      ret = True
-      break
+      builder.table_ea = refs[0].ea
+      builder.offset = 0
+      builder.offset_mult = 1
+
+      # Don't treat things like thunks to be tables.
+      if is_thunk(builder.table_ea) or is_external_segment(builder.table_ea):
+        builder.table_ea = idc.BADADDR
+        continue
+
+      if try_get_simple_jump_table_reader(builder):
+        ret = True
+        break
   
   if ret:
     DEBUG("Reader inferred jump table base: {:x}".format(builder.table_ea))
@@ -436,12 +520,52 @@ def get_manual_jump_table_reader(builder):
   #            based jump tables works. 
   return False
 
+def get_dref_jump_table_reader(builder):
+  """Try to get a jump table by looking at any of the blocks that might be
+  referenced by data."""
+  func = idaapi.get_func(builder.jump_ea)
+  if not func:
+    return False
+
+  min_entry_ea = idc.BADADDR
+  max_entry_ea = 0
+  
+  import flow
+
+  best_entry_ea = 0
+  best_num_entries = 0
+
+  for block_ea in flow.find_default_block_heads(func.start_ea):
+    for entry_ea in drefs_to(block_ea):
+      builder.table_ea = entry_ea
+      builder.offset = 0
+      builder.offset_mult = 1
+      del builder.candidate_target_eas[:]
+      if try_get_simple_jump_table_reader(builder):
+        if len(builder.candidate_target_eas) > best_num_entries:
+          best_entry_ea = entry_ea
+
+  if best_num_entries:
+    builder.table_ea = entry_ea
+    builder.offset = 0
+    builder.offset_mult = 1
+    del builder.candidate_target_eas[:]
+    return try_get_simple_jump_table_reader(builder)
+  
+  return False
+
+
 def get_jump_table_reader(builder):
   """Returns the size of a jump table entry, as well as a reader function
   that can extract entries."""
   si = ida_nalt.get_switch_info(builder.jump_ea)
   if si:
-    return get_ida_jump_table_reader(builder, si)
+    if get_ida_jump_table_reader(builder, si):
+      return True
+    else:
+      builder.offset = 0
+      builder.offset_mult = 1
+      del builder.candidate_target_eas[:]
 
   # IDA can be a bit ignorant at recognizing jump tables. This came up
   # in sqlite3 where IDA decided that `jmp ds:off_48A5F0[rax*8]` wasn't
@@ -450,39 +574,20 @@ def get_jump_table_reader(builder):
   # opposed to being an `o_disp`. `get_instruction_references` correctly
   # resolves this difference, so we'll also try to use it to pick up
   # where IDA leaves off.
-  else:
-    return get_manual_jump_table_reader(builder)
+  if get_manual_jump_table_reader(builder):
+    return True
+
+  return False
+  #return get_dref_jump_table_reader(builder)
 
 _JMP_THROUGH_TABLE_INFO = {}
+_TABLE_INFO = {}
 _NOT_A_JMP_THROUGH_TABLE = set()
 
-def get_jump_table(inst, binary_is_pie=False):
-  """Returns an instance of JumpTable, or None depending on whether or not
-  a jump table was discovered."""
-  global _JMP_THROUGH_TABLE_INFO, _NOT_A_JMP_THROUGH_TABLE
-  global _INVALID_JMP_TABLE
-
-  if not inst or not is_indirect_jump(inst):
-    return None  # Don't cache.
-
-  if inst.ea in _JMP_THROUGH_TABLE_INFO:
-    return _JMP_THROUGH_TABLE_INFO[inst.ea]
-
-  elif inst.ea in _NOT_A_JMP_THROUGH_TABLE:
-    return None
-
-  elif is_external_segment(inst.ea):
-    _NOT_A_JMP_THROUGH_TABLE.add(inst.ea)
-    return None
-
-  builder = JumpTableBuilder(inst, binary_is_pie)
-
-  if not get_jump_table_reader(builder):
-    _NOT_A_JMP_THROUGH_TABLE.add(builder.jump_ea)
-    return None
-
-  DEBUG("Jump table candidate at {:x} referenced by instruction {:x}".format(
-      builder.table_ea, builder.jump_ea))
+def _handle_new_builder(builder):
+  """Try to finalize a new jump table builder and get the final jump
+  table."""
+  global _NOT_A_JMP_THROUGH_TABLE, _IS_TARGETED_BY_JUMP_TABLE
 
   get_default_jump_table_entries(builder)
 
@@ -506,19 +611,62 @@ def get_jump_table(inst, binary_is_pie=False):
   for i in xrange(num_entries):
     entry_data = builder.read_entry(entry_addr)
     entries[entry_addr] = entry_data
+    _IS_TARGETED_BY_JUMP_TABLE.add(entry_data)
     DEBUG("  {:x} => {:x}".format(entry_addr, entry_data))
     entry_addr += builder.entry_size * builder.entry_mult
 
   table = JumpTable(builder, entries)
   _JMP_THROUGH_TABLE_INFO[builder.jump_ea] = table
+  _TABLE_INFO[builder.table_ea] = table
 
   return table
+
+def get_jump_table(inst, binary_is_pie=False):
+  """Returns an instance of JumpTable, or None depending on whether or not
+  a jump table was discovered."""
+  global _JMP_THROUGH_TABLE_INFO, _NOT_A_JMP_THROUGH_TABLE, _TABLE_INFO
+  global _INVALID_JMP_TABLE, _IS_TARGETED_BY_JUMP_TABLE
+
+  if not inst or not is_indirect_jump(inst):
+    return None  # Don't cache.
+
+  if inst.ea in _JMP_THROUGH_TABLE_INFO:
+    return _JMP_THROUGH_TABLE_INFO[inst.ea]
+
+  elif inst.ea in _NOT_A_JMP_THROUGH_TABLE:
+    return None
+
+  elif is_external_segment(inst.ea):
+    _NOT_A_JMP_THROUGH_TABLE.add(inst.ea)
+    return None
+
+  builder = JumpTableBuilder(inst, binary_is_pie)
+
+  if not get_jump_table_reader(builder):
+    _NOT_A_JMP_THROUGH_TABLE.add(builder.jump_ea)
+    return None
+
+  DEBUG("Jump table candidate at {:x} referenced by instruction {:x}".format(
+      builder.table_ea, builder.jump_ea))
+
+  if builder.table_ea in _TABLE_INFO:
+    DEBUG("  Using pre-existing jump table info")
+    table = _TABLE_INFO[builder.table_ea]
+    _JMP_THROUGH_TABLE_INFO[builder.jump_ea] = table
+    return table
+
+  return _handle_new_builder(builder)
 
 def is_jump_table_entry(ea):
   """Returns `True` if `ea` falls somewhere inside of the bytes of a jump
   table."""
-  global _FIRST_JUMP_TABLE_ENTRY
-  return ea in _FIRST_JUMP_TABLE_ENTRY
+  global _IS_JUMP_TABLE_ENTRY
+  return ea in _IS_JUMP_TABLE_ENTRY
+
+def is_jump_table_target(ea):
+  """Returns `True` if `ea` is targeted by some jump table."""
+  global _IS_TARGETED_BY_JUMP_TABLE
+  return ea in _IS_TARGETED_BY_JUMP_TABLE
 
 def get_jump_table_from_entry(entry_ea):
   """Returns a `JumpTable` """
@@ -527,3 +675,60 @@ def get_jump_table_from_entry(entry_ea):
     return None
   table_ea = _FIRST_JUMP_TABLE_ENTRY[entry_ea]
   return _JUMP_TABLE_ENTRY[table_ea]
+
+def _find_jumps_near(inst_ea):
+  """Try to find an indirect jump instruction near `inst_ea`."""
+  min_ea, max_ea = get_function_bounds(inst_ea)
+  candidates = []
+  DEBUG("Looking for indirect jumps in range [{:x}, {:x})".format(min_ea, max_ea))
+  while min_ea < max_ea:
+    if is_indirect_jump(min_ea):
+      DEBUG("  Found indirect jump at {:x}".format(min_ea))
+      candidates.append(min_ea)
+    min_ea = idc.next_head(min_ea+1)
+  return candidates
+
+def try_create_jump_table(inst_ea, entry_ea, entry_size, binary_is_pie=False):
+  """Try to create a jump table, beginning at `entry_ea`, and referenced by
+  `inst_ea` (which may or may not be a jump instruction)."""
+  DEBUG_PUSH()
+
+  if entry_ea in _TABLE_INFO:
+    table = _TABLE_INFO[entry_ea]
+    DEBUG("Instruction at {:x} references jump table {:x} because of entry at {:x}".format(
+        inst_ea, table.table_ea, entry_ea))
+    DEBUG_POP()
+    return
+
+  if not is_indirect_jump(inst_ea):
+    jump_eas = _find_jumps_near(inst_ea)
+  else:
+    jump_eas = [inst_ea]
+
+  for jump_ea in jump_eas:
+    if jump_ea in _JMP_THROUGH_TABLE_INFO:
+      continue  # Already have a jump table for this.
+
+    # elif entry_ea in _TABLE_INFO:
+    #   table = _TABLE_INFO[entry_ea]
+    #   DEBUG("Jump table candidate at {:x} referenced by instruction {:x} because of entry {:x}".format(
+    #       table.table_ea, jump_ea, entry_ea))
+    #   _JMP_THROUGH_TABLE_INFO[jump_ea] = table
+    #   continue
+
+    inst, _ = decode_instruction(jump_ea)
+    builder = JumpTableBuilder(inst, binary_is_pie)
+    builder.table_ea = entry_ea
+    builder.entry_size = entry_size
+
+    if not try_get_simple_jump_table_reader(builder):
+      continue
+
+    DEBUG("Jump table candidate at {:x} referenced by instruction {:x}".format(
+        builder.table_ea, builder.jump_ea))
+
+    _handle_new_builder(builder)
+    break
+
+  DEBUG_POP()
+
